+++ instantwords/lib/core/authentication/login.dart
@@ -1,4 +1,4 @@
-part of '../main.dart';
+part of '../../main.dart';
 
 class LoginPage extends StatefulWidget {
   final FireStorage _storage;
