+++ instantwords/lib/core/authentication/register.dart
@@ -1,4 +1,4 @@
-part of '../main.dart';
+part of '../../main.dart';
 
 class RegisterPage extends StatefulWidget {
   final FireStorage _storage;
