+++ instantwords/lib/core/database/authexceptionhandler.dart
@@ -1,0 +1,87 @@
+part of '../../main.dart';
+
+enum AuthResultStatus {
+  successful,
+  emailAlreadyExists,
+  wrongPassword,
+  invalidEmail,
+  userNotFound,
+  userDisabled,
+  operationNotAllowed,
+  tooManyRequests,
+  undefined,
+  weakPassword
+}
+
+class AuthExceptionHandler {
+  static handleException(e) {
+    var status;
+    switch (e.code) {
+      case "invalid-email":
+        status = AuthResultStatus.invalidEmail;
+        break;
+      case "wrong-password":
+        status = AuthResultStatus.wrongPassword;
+        break;
+      case "user-not-found":
+        status = AuthResultStatus.userNotFound;
+        break;
+      case "ERROR_USER_DISABLED":
+        status = AuthResultStatus.userDisabled;
+        break;
+      case "ERROR_TOO_MANY_REQUESTS":
+        status = AuthResultStatus.tooManyRequests;
+        break;
+      case "operation-not-allowed":
+        status = AuthResultStatus.operationNotAllowed;
+        break;
+      case "email-already-in-use":
+        status = AuthResultStatus.emailAlreadyExists;
+        break;
+      case "weak-password":
+        status = AuthResultStatus.weakPassword;
+        break;
+      default:
+        status = AuthResultStatus.undefined;
+    }
+    return status;
+  }
+
+  ///
+  /// Accepts AuthExceptionHandler.errorType
+  ///
+  static generateExceptionMessage(exceptionCode) {
+    String errorMessage;
+    switch (exceptionCode) {
+      case AuthResultStatus.invalidEmail:
+        errorMessage = "Your email address appears to be malformed.";
+        break;
+      case AuthResultStatus.wrongPassword:
+        errorMessage = "Your password is wrong.";
+        break;
+      case AuthResultStatus.userNotFound:
+        errorMessage = "User with this email doesn't exist.";
+        break;
+      case AuthResultStatus.userDisabled:
+        errorMessage = "User with this email has been disabled.";
+        break;
+      case AuthResultStatus.tooManyRequests:
+        errorMessage = "Too many requests. Try again later.";
+        break;
+      case AuthResultStatus.operationNotAllowed:
+        errorMessage = "Signing in with Email and Password is not enabled.";
+        break;
+      case AuthResultStatus.emailAlreadyExists:
+        errorMessage =
+            "The email has already been registered. Please login or reset your password.";
+        break;
+      case AuthResultStatus.weakPassword:
+        errorMessage = "Password should be at least 6 characters.";
+        break;
+      default:
+        errorMessage = "An undefined Error happened.";
+    }
+
+    return errorMessage;
+  }
+}
